requirements.txt
discord.py==1.7.3
Pillow==8.3.1
urllib3==1.26.6
certifi==2021.5.30
charset-normalizer==2.0.4
requests==2.26.0
